Remove --no-deamon from gradle invocations (#1)
Gradle will always spin up a deamon since the non daemon mode got remove couple major versions ago. With that flag gradle will though shutdown the daemon at the end of the execution which incurs extra overhead in case of this action as it runs gradle 3 times. Without the flag we will spin up a daemon on first run and latter two will get to reuse it